requirements.txt
# altair
# pandas
streamlit==1.35.0
pymysql==1.1.1
SQLAlchemy==2.0.30
openai==1.95.1
python-dotenv==1.0.1
pandas==2.2.2
# faiss-cpu==1.7.4
# langchain==0.2.0
chromadb
tqdm
seaborn
matplotlib
plotly
langchain==0.1.16
langchain-openai==0.1.6
langchain-experimental==0.0.52
